Optimize Request Service and Repository

+++ src/main/java/com/it120p/carehub/service/RequestService.java
@@ -1,5 +1,7 @@
 package com.it120p.carehub.service;
 
+import com.it120p.carehub.exceptions.MissingException;
+import com.it120p.carehub.exceptions.PermissionException;
 import com.it120p.carehub.model.entity.Request;
 import com.it120p.carehub.model.entity.User;
 import com.it120p.carehub.repository.RequestRepository;
@@ -8,6 +10,7 @@ import org.springframework.data.util.Streamable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RequestService {
@@ -27,13 +30,34 @@ public class RequestService {
         return requestRepository.findRequestsByUser(customer);
     }
 
-    public Request getRequestFromUser(User customer, long requestId) {
-        return requestRepository.findRequestByUserIdAndRequestId(customer, requestId).orElseThrow();
+    public Request getOwnedRequest(User customer, long requestId) throws Exception {
+        // Get Request
+        Optional<Request> optionalRequest = requestRepository.findRequestById(requestId);
+        if (optionalRequest.isEmpty()) throw new MissingException("Request");
+
+        Request request = optionalRequest.get();
+
+        // Check if owner
+        if (request.getCustomer().getUserId() != customer.getUserId()) throw new PermissionException("Request");
+
+        // Return request
+        return request;
     }
 
-    public Request removeRequestFromUser(User customer, long requestId) {
-        Request request = requestRepository.findRequestByUserIdAndRequestId(customer, requestId).orElseThrow();
+    public Request removeOwnedRequest(User customer, long requestId) throws Exception {
+        // Get Request
+        Optional<Request> optionalRequest = requestRepository.findRequestById(requestId);
+        if (optionalRequest.isEmpty()) throw new MissingException("Request");
+
+        Request request = optionalRequest.get();
+
+        // Check if owner
+        if (request.getCustomer().getUserId() != customer.getUserId()) throw new PermissionException("Request");
+
+        // Remove request
         requestRepository.delete(request);
+
+        // Return request
         return request;
     }
 }
